Fix token authorizer bugs and add new authorizer
This commit fixes bugs in the API token authorizer and adds a new
authorizer class.

The changes include:

- Fixed a bug in the `Authorizer` class where the API token authorizer
was not being set properly.
- Added a new authorizer class called `LoggedOnWithShortLivedAuthorizer`
that checks if the user is logged on with a short-lived token.
- Updated the `APITokenAuthorizer` class to properly set the authorizer
object if provided.
- Added a new exception class `APITokenAuthorizerAlreadySetException`
to handle cases where the API token authorizer is already set.
- Updated the `logout` function to use the
`LoggedOnWithShortLivedAuthorizer` authorizer.
- Added a new test case to test logging out with a long-lived token.

These changes improve the authentication functionality of the API.