Create class to generate resource CRUD endpoints
I have created the `ResourceCRUDAPIRouter` class that can be used to
generate a FastAPI `APIRouter` that contains the CRUD operations for
specific endpoints. By doing this, we can generate endpoint dynamically
with specific given parameters. This way, we don't have to copy and
paste code and we don't have to repeat outselves. I see this as a
absolute win.