Refactor resource endpoints
We were working on making a class that could automatically generate
endpoints for resources. This was working from a code point-of-view, but
was very bad for FastAPI. Because we were adding request values and
return types based on a `Generic`, FastAPI couldn't figure out what
kind of models were expected. This resulted in very unpredictable
behaviour.

We modified this class to perform the correct code to retrieve objects
from the `MyData` backend and we created _normal_ FastAPI endpoints
to retrieve the data that delegate the work to this class. This way, we
can leverage the use of FastAPI and still make sure we don't have
duplicate code.